Revert "Update gitlab4j to 5.x (#976)" (#979)
This reverts commit 15faa6153560173c0ab7f1fdd37136e1b6caa5f2.

+++ legend-sdlc-server/src/main/java/org/finos/legend/sdlc/server/gitlab/api/GitLabProjectApi.java
@@ -283,7 +283,7 @@ public class GitLabProjectApi extends GitLabApiWithFileAccess implements Project
 
     private boolean tryEnsureMinAccessLevel(GitLabApi gitLabApi, org.gitlab4j.api.models.Project project, AccessLevel accessLevel)
     {
-        long projectId = project.getId();
+        int projectId = project.getId();
 
         User currentUser;
         try
@@ -298,7 +298,7 @@ public class GitLabProjectApi extends GitLabApiWithFileAccess implements Project
         }
 
         org.gitlab4j.api.ProjectApi projectApi = gitLabApi.getProjectApi();
-        long userId = currentUser.getId();
+        int userId = currentUser.getId();
         CallUntil<AccessLevel, ?> callUntil = CallUntil.callUntil(() ->
         {
             try
